remove duplicate events
after the events list has been sorted, the loops goes through and checks to see if the event is the same as the previous one. if it is, that event is deleted. note: this does not delete repeating events, only events where the time, date, and event name are exactly the same. For some reason I had a lot of events that were duplicated.